+++ src/plugin/integration/SocketConnection.cpp
@@ -1,17 +1,17 @@
-#include "pch/pch.h"
-#include "integration/SocketConnection.h"
-#include "integration/SocketInterface.h"
+#include "SocketConnection.h"
+#include "SocketInterface.h"
 
 namespace UKControllerPlugin::Integration {
-    SocketConnection::SocketConnection(std::shared_ptr<SocketInterface> socket): socket(socket)
-    { }
+    SocketConnection::SocketConnection(std::shared_ptr<SocketInterface> socket) : socket(std::move(socket))
+    {
+    }
 
-    bool SocketConnection::Active() const
+    auto SocketConnection::Active() const -> bool
     {
         return this->socket->Active();
     }
 
-    std::queue<std::string> SocketConnection::Receive()
+    auto SocketConnection::Receive() -> std::queue<std::string>
     {
         *this->socket >> this->incomingData;
         return this->ProcessReceivedData();
@@ -23,27 +23,25 @@ namespace UKControllerPlugin::Integration {
         *this->socket << message;
     }
 
-    std::queue<std::string> SocketConnection::ProcessReceivedData()
+    auto SocketConnection::ProcessReceivedData() -> std::queue<std::string>
     {
         if (this->incomingData.tellp() == 0) {
             return {};
         }
 
-        std::string streamData;
+        std::string streamData = this->incomingData.str();
         std::queue<std::string> messages;
-        this->incomingData >> streamData;
 
-        size_t indexOfMessageSeparator;
+        size_t indexOfMessageSeparator = 0;
         while ((indexOfMessageSeparator = streamData.find(MESSAGE_DELIMITER)) != std::string::npos) {
             messages.push(streamData.substr(0, indexOfMessageSeparator));
             streamData.erase(0, indexOfMessageSeparator + 1);
         }
 
         // Put any incomplete messages back on the stream for later
-        if (!streamData.empty()) {
-            this->incomingData = std::stringstream();
-            this->incomingData << streamData;
-        }
+        this->incomingData.str("");
+        this->incomingData.clear();
+        this->incomingData << streamData;
 
         return messages;
     }