+++ src/plugin/releases/DepartureReleaseDecisionList.h
@@ -71,6 +71,11 @@ namespace UKControllerPlugin::Releases {
         const Gdiplus::Rect airportColumnHeader{225, 5, 40, 25};
         const Gdiplus::Rect sidColumnHeader{275, 5, 65, 25};
 
+        // Some colours
+        const Gdiplus::Color OFF_WHITE_COLOUR = Gdiplus::Color(255, 255, 255);
+        const Gdiplus::Color TITLE_BAR_BASE_COLOUR = Gdiplus::Color(130, 50, 154);
+        const Gdiplus::Color TITLE_BAR_FLASH_COLOUR = Gdiplus::Color(179, 3, 0);
+
         // Brushes
         const Gdiplus::SolidBrush textBrush;
 
@@ -102,11 +107,6 @@ namespace UKControllerPlugin::Releases {
 
         // Our window position
         Gdiplus::PointF position = {0, 0};
-
-        // Some colours
-        const Gdiplus::Color OFF_WHITE_COLOUR = Gdiplus::Color(227, 227, 227);
-        const Gdiplus::Color TITLE_BAR_BASE_COLOUR = Gdiplus::Color(197, 129, 214);
-        const Gdiplus::Color TITLE_BAR_FLASH_COLOUR = Gdiplus::Color(255, 97, 93);
 
         // Standard window components
         std::shared_ptr<Components::TitleBar> titleBar;