+++ docs/UserGuide/Changelog/Changelog.md
@@ -1,5 +1,27 @@
 # UK Controller Plugin Changelog
 
+# [3.7.0](https://github.com/VATSIM-UK/uk-controller-plugin/compare/3.6.3...3.7.0) (2021-10-26)
+
+
+### Features
+
+* **selcal:** add tag item to show selcal from flightplan remarks ([#353](https://github.com/VATSIM-UK/uk-controller-plugin/issues/353)) ([c26dc46](https://github.com/VATSIM-UK/uk-controller-plugin/commit/c26dc46eca5c9f8ab71322cd5a202481b7480e76))
+
+## [3.6.3](https://github.com/VATSIM-UK/uk-controller-plugin/compare/3.6.2...3.6.3) (2021-10-18)
+
+
+### Bug Fixes
+
+* **release+prenotes:** Broken colours in lists ([#354](https://github.com/VATSIM-UK/uk-controller-plugin/issues/354)) ([292a613](https://github.com/VATSIM-UK/uk-controller-plugin/commit/292a6134191006e602ae2d5f5e7c0960b66bae7e))
+
+## [3.6.2](https://github.com/VATSIM-UK/uk-controller-plugin/compare/3.6.1...3.6.2) (2021-10-02)
+
+
+### Bug Fixes
+
+* **a11y:** Make release and prenote list header colours accessible ([e89dd19](https://github.com/VATSIM-UK/uk-controller-plugin/commit/e89dd1929d4d277662af71549e3a52879a69318d))
+* **integration:** Fix messages not being received from clients ([#343](https://github.com/VATSIM-UK/uk-controller-plugin/issues/343)) ([7d85c9b](https://github.com/VATSIM-UK/uk-controller-plugin/commit/7d85c9b5401dcd28e29b999cb4de904cf4d28b99))
+
 ## [3.6.1](https://github.com/VATSIM-UK/uk-controller-plugin/compare/3.6.0...3.6.1) (2021-09-18)